+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unpublished
+### 🚀 Features
+- AndesProgressIndicatorIndeterminate: New component AndesProgressIndicator| Authors: [@javff](https://github.com/javff)
+
 ### 🛠 Bug fixes
 - AndestextField: Fix icon component right and left interaction | Authors: [@joalonsopint](https://github.com/joalonsopint)